pyproject.toml
[project]
name = "doctown-v12"
version = "0.1.0"
description = "Intent-driven document analysis and structured extraction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.6.1",
    "requests>=2.32.5",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]
